feat: add a no-negated-async rule
Angular's async pipes emit `null` initially, prior to the observable emitting
any values, or the promise resolving. This can cause negations, like
`*ngIf="!(myConditional | async)"` to thrash the layout and cause expensive
side-effects like firing off XHR requests for a component which should not
be shown.

This PR adds a linting rule which fails on usages of !(expr | async) and
(expr | async) == false.